+++ zbar/decoder/code128.c
@@ -256,7 +256,7 @@ static inline signed char decode6(zbar_decoder_t *dcode)
     bars = bars * 11 * 4 / s;
     chk	 = calc_check(c);
     dbprintf(2, " bars=%d chk=%d", bars, chk);
-    if (chk - 7 > bars || bars > chk + 7)
+    if ((int)(chk - 7) > (int)bars || (int)bars > (int)(chk + 7))
 	return (-1);
 
     return (c & 0x7f);
@@ -366,7 +366,7 @@ static inline unsigned char postprocess(zbar_decoder_t *dcode)
     if (dcode128->direction) {
 	/* reverse buffer */
 	dbprintf(2, " (rev)");
-	for (i = 0; i < dcode128->character / 2; i++) {
+	for (i = 0; (int)i < dcode128->character / 2; i++) {
 	    unsigned j	  = dcode128->character - 1 - i;
 	    code	  = dcode->buf[i];
 	    dcode->buf[i] = dcode->buf[j];
@@ -388,7 +388,7 @@ static inline unsigned char postprocess(zbar_decoder_t *dcode)
     cexp    = (code == START_C) ? 1 : 0;
     dbprintf(2, " start=%c", 'A' + charset);
 
-    for (i = 1, j = 0; i < dcode128->character - 2; i++) {
+    for (i = 1, j = 0; (int)i < dcode128->character - 2; i++) {
 	unsigned char code = dcode->buf[i];
 	zassert(!(code & 0x80), 1,
 		"i=%x j=%x code=%02x charset=%x cexp=%x %s\n", i, j, code,
@@ -435,7 +435,7 @@ static inline unsigned char postprocess(zbar_decoder_t *dcode)
 		    dcode->modifiers |= MOD(ZBAR_MOD_GS1);
 		else if (i == 2)
 		    dcode->modifiers |= MOD(ZBAR_MOD_AIM);
-		else if (i < dcode->code128.character - 3)
+		else if ((int)i < dcode->code128.character - 3)
 		    dcode->buf[j++] = 0x1d;
 		/*else drop trailing FNC1 */
 	    } else if (code >= START_A) {
@@ -541,7 +541,7 @@ zbar_symbol_type_t _zbar_decode_code128(zbar_decoder_t *dcode)
     }
     dcode128->width = dcode128->s6;
 
-    zassert(dcode->buf_alloc > dcode128->character, 0,
+    zassert((int)dcode->buf_alloc > dcode128->character, 0,
 	    "alloc=%x idx=%x c=%02x %s\n", dcode->buf_alloc,
 	    dcode128->character, c,
 	    _zbar_decoder_buf_dump(dcode->buf, dcode->buf_alloc));
